Try to fit with mastodon hierarchy.
There are a lot of copy/paste by just injecting some specific
classes for the junction graph. The main reason is that everything
done in the bdv.overlay.wrap package of mastodon is specific to
Spot. Actually, not the Spot class, but vertices that have a label
and a covariance.